-Added Sword object, that is drawn at the top of the player's position, I tried with different direction but it caused some issues so that has to be sorted
-for the player, map objects, and sword the collision boxes are drawn properly based on vertex count, though the squares have a weird visual glitch they do draw 4 edges

-the sword object itself is hidden and not used until you press and hold the home button (will change later) which should draw the collision box (animated sprite is not setup)
afterwards any breakable object you collide into while it has been drawn will have it's collision box basically turned off and not drawn, which then can be later added to be removed from the map

-Player, Map object, and Sword have been setup with enums that show what object they are, which is used for the collision listener, you basically determine what they are and do various actions according
So now when a player collides with a breakable object it will not do what the sword does and destroy the object. (took awhile to find a fix for this), this can be setup for projectiles, enemies and any other entities, since the box2d stores memory about
the object, the constructor needs to be setup to have the class type as blank.

Things to do:
-Animated Sprite needs fixing
-Sword Direction needs fixing

Bugs
-Visual Bug with debug draw for collisions
-If the sword is being used (button held down) but is inside a breakable object it does not currently think it is colliding, will have to look into fixes for that